tests: reorganize test modules
- Rename `test_plugins_input` to `test_plugin_userinput`
- Move CLI tests to tests/cli and rename test modules accordingly
- Fix relative module import paths
- Fix logging handler being added to the root logger, causing issues
  with other tests, depending on the order of test execution
