+++ tests/cli/output/test_playeroutput.py
