+++ tests/cli/test_cmdline_player_fifo.py
@@ -1,7 +1,7 @@
 from unittest.mock import Mock, patch
 
 from tests import posix_only, windows_only
-from tests.test_cmdline import CommandLineTestCase
+from tests.cli.test_cmdline import CommandLineTestCase
 
 
 @posix_only
