+++ tests/cli/test_main_formatter.py
